Separate GroupListView form SessionProfileWindow
