+++ SessionProfileWindowController.swift
@@ -9,19 +9,48 @@
 import Cocoa
 
 class SessionProfileWindowController: NSWindowController {
-    var object: SessionProfileController!
+    @IBOutlet weak var groupListView:GroupListView!
+    @IBOutlet weak var formView:FormView!
     
-    convenience init() {
-        let object = SessionProfileController()
-        self.init(windowNibName: "SessionProfileWindow", owner: object)
-        self.object = object
+    var profile: SessionProfile!
+    
+    convenience init(profile: SessionProfile) {
+        self.init(windowNibName: "SessionProfileWindow")
+        self.profile = profile
     }
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        object.sourceListView!.expandItem(nil, expandChildren: true)
-        object.sourceListView!.selectRowIndexes(NSIndexSet(index: 0), byExtendingSelection: false)
-        object.window!.title = "New Session profile"
-        self.object.window.makeKeyAndOrderFront(self)
+        self.groupListView.reloadData()
+        self.groupListView.expandItem(nil, expandChildren: true)
+        self.groupListView.selectRowIndexes(NSIndexSet(index: 0), byExtendingSelection: false)
+        self.window!.title = "New Session profile"
+        self.window!.makeKeyAndOrderFront(self)
+        
+        self.updateWindowTitle()
+    }
+    
+    private func updateWindowTitle() {
+        if self.window == nil {
+            return
+        }
+        
+        if self.profile.name == "" {
+            self.window!.title = "New Session profile"
+        } else {
+            self.window!.title = self.profile.name
+        }
+    }
+    
+    static var fieldsXML: XMLIndexer?
+    var fields: XMLElement {
+        if SessionProfileWindowController.fieldsXML == nil {
+            let filePath: NSString = NSBundle.mainBundle().pathForResource("SessionProfileFields", ofType: "xml")!
+            if let data: NSData = try! NSData(contentsOfFile: String(filePath), options: []) {
+                SessionProfileWindowController.fieldsXML = SWXMLHash.parse(data)
+            }
+        }
+        
+        return SessionProfileWindowController.fieldsXML!["property"].element!
     }
 }
